move DevRelease enum to avoid akward imports
