+++ src/daxxl/defs.py
@@ -229,3 +229,7 @@ class ModEntry:
         self.new_contributors: List[str] = []
         self.oldest_link_version = ""
         self.newest_link_version = ""
+
+class DevRelease(str, Enum):
+    DAILY = "daily"
+    EXPERIMENTAL = "experimental"
